Gate delete done AL

+++ app/Http/Controllers/ArticleController.php
@@ -99,6 +99,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        $this->authorize('canDelete', $article);
         $article->delete();
         return redirect()->back();
     }
